app.tsからline関連処理を切り出す
